TF-33

+++ app/Http/Controllers/TrackController.php
@@ -11,24 +11,36 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
+use Symfony\Component\HttpFoundation\Response;
 
 class TrackController extends Controller
 {
     public function index(TrackService $service): AnonymousResourceCollection
     {
+        Gate::authorize('viewAny', Track::class);
+
         return TrackResource::collection($service->get($this->user));
     }
 
     public function getActive(TrackService $service): TrackResource | JsonResponse
     {
-        if ($track = $service->getActive($this->user)) return TrackResource::make($track);
+        $track = $service->getActive($this->user);
+
+        if ($track) {
+            Gate::authorize('view', $track);
+
+            return TrackResource::make($track);
+        }
 
         return response()->json(['data' => null]);
     }
 
-    public function store(Request $request, TrackService $service): TrackResource
+    public function store(Request $request, TrackService $service): JsonResponse
     {
+        Gate::authorize('create', Track::class);
+
         $request->validate([
             'title' => ['nullable', 'string', 'max:1000'],
             'started_at' => ['nullable', Rule::date()->format('Y-m-d H:i:s')],
@@ -44,11 +56,15 @@ class TrackController extends Controller
             )
         );
 
-        return TrackResource::make($track);
+        return TrackResource::make($track)
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);
     }
 
-    public function update(Request $request, Track $track, TrackService $service): TrackResource
+    public function update(Request $request, Track $track, TrackService $service): JsonResponse
     {
+        Gate::authorize('update', $track);
+
         $request->validate([
             'title' => ['nullable', 'string', 'max:1000'],
             'started_at' => ['nullable', Rule::date()->format('Y-m-d H:i:s')],
@@ -64,11 +80,13 @@ class TrackController extends Controller
             )
         );
 
-        return TrackResource::make($track);
+        return TrackResource::make($track)->response();
     }
 
     public function destroy(Track $track, TrackService $service): JsonResponse
     {
+        Gate::authorize('delete', $track);
+
         $service->delete($track);
 
         return response()->json();
